+++ CHANGES.md
@@ -10,6 +10,7 @@
 * fixed [#714](https://github.com/afragen/github-updater/issues/714), get correct Bitbucket release asset download link from AWS
 * update to `class-parser.php` r7679
 * don't run on heartbeat API 💗
+* only run on `admin-ajax.php` when possibly attempting sequential shiny updates, fixes [#723](https://github.com/afragen/github-updater/issues/723)
 
 #### 8.3.1 / 2018-09-13
 * created `class Bootstrap` to setup plugin loading